backend/app/requirements.txt
fastapi
sqlalchemy
python-dotenv
psycopg2-binary
boto3
requests
uvicorn[standard]
httpx
gunicorn
pandas
scipy